+++ pysrc/preprocess/embeddings/embeddings_model_connector.py
@@ -18,7 +18,8 @@ class EmbeddingsModelConnector:
 
         self.text_embedding = lambda t: self.sentence_transformer_model.encode(t)
         self.batch_texts_embeddings = lambda t: self.sentence_transformer_model.encode(t)
-        self.embeddings_model_name = re.sub('[^a-zA-Z0-9]+', '_', self.sentence_transformer_model.model_name)
         self.device = self.sentence_transformer_model.device
         emb = self.sentence_transformer_model.encode(['This is a test.', 'This is a test2'])
-        self.embeddings_dimension = emb.shape[1]
+        dim = emb.shape[1]
+        assert dim == config.embeddings_dimension, \
+            f'Unexpected embeddings dimension {dim} != {config.embeddings_dimension}'