+++ pysrc/services/embeddings_service.py
@@ -4,10 +4,14 @@ import requests
 
 import numpy as np
 
+from pysrc.preprocess.embeddings.embeddings_db_connector import EmbeddingsDBConnector
+
 logger = logging.getLogger(__name__)
 
 # Launch with a Docker address or locally
 EMBEDDINGS_SERVICE_URL = os.getenv('EMBEDDINGS_SERVICE_URL', 'http://localhost:5001')
+
+EMBEDDINGS_DB_CONNECTOR = EmbeddingsDBConnector()
 
 def is_embeddings_service_available():
     logger.debug(f'Check if embeddings service endpoint is available')
@@ -32,7 +36,6 @@ def is_embeddings_service_ready():
     except Exception as e:
         logger.debug(f'Embeddings service is not ready: {e}')
         return False
-
 
 def fetch_tokens_embeddings(tokens):
     # Don't use the model as is, since each celery process will load its own copy.
@@ -76,3 +79,13 @@ def fetch_texts_embedding(texts):
     except Exception as e:
         logger.debug(f'Failed to fetch texts embeddings ${e}')
     return None
+
+
+def is_embeddings_db_available():
+    logger.debug(f'Check if embeddings db is available')
+    return EMBEDDINGS_DB_CONNECTOR.is_embeddings_db_available()
+
+def load_embeddings_from_df(pids):
+    logger.debug("Fetching embeddings from DB")
+    index, embeddings = EMBEDDINGS_DB_CONNECTOR.load_embeddings_by_ids(pids)
+    return np.array(embeddings).reshape(len(index), -1), [(str(p), c) for p, c in index]