+++ CHANGES.md
@@ -3,9 +3,16 @@ Pubtrends changelog
 
 Here you can see the full list of changes between each release.
 
+Version 1.4
+------------
+Released on Aug XX, 2025
+
+- Reuse embeddings DB for fetching text embeddings
+- Use parquet format for Faiss index storage
+
 Version 1.3
 ------------
-Released on XXX XX, 2025
+Released on Jun 01, 2025
 
 - Reworked text embeddings usage for global similarity network construction
 - Supported Sentence Transformer models for whole texts embeddings